apps/server-2/requirements.txt
# python-opencv-headless==4.11.0.86  # safer for servers, no GUI deps
# pytesseract==0.3.13
# numpy>=1.25.0
# pdfplumber>=0.11.0
# Pillow>=10.0.0
# fastapi>=0.100.0
# uvicorn>=0.23.0
# python-multipart>=0.0.6 # if you later handle file uploads in FastAPI

opencv-python-headless==4.11.0.86  # safer for servers, no GUI deps
pytesseract==0.3.13
numpy>=1.25.0
pdfplumber>=0.11.0
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
# python-multipart>=0.0.6# if you later handle file uploads in FastAPI